Separating the facade class out into its own class
