+++ src/main/java/com/telenordigital/prime/firebase/FbStorage.java
@@ -6,9 +6,7 @@ import com.google.firebase.auth.FirebaseCredentials;
 import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 import com.telenordigital.prime.events.Product;
 import com.telenordigital.prime.events.ProductDescriptionCache;
@@ -26,13 +24,9 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -41,278 +35,6 @@ public final class FbStorage implements Storage {
     private static final Logger LOG = LoggerFactory.getLogger(FbStorage.class);
 
     private final ProductDescriptionCache productCache;
-
-
-    /**
-     * Presenting a fascade for the many and  varied firebase databases
-     * we're using.
-     */
-    public  final static class FbDatabaseFacade {
-
-        private final DatabaseReference authorativeUserData;
-        private final DatabaseReference clientRequests;
-        private final DatabaseReference clientVisibleSubscriberRecords;
-        private final DatabaseReference recordsOfPurchase;
-        private final DatabaseReference quickBuyProducts;
-        private final DatabaseReference products;
-
-        FbDatabaseFacade(FirebaseDatabase firebaseDatabase) {
-            checkNotNull(firebaseDatabase);
-            // XXX Read this, then fix something that reports connectivity status through the
-            //     health mechanism.
-            // https://www.firebase.com/docs/web/guide/offline-capabilities.html#section-connection-state
-            // this.firebaseDatabase.getReference("/.info/connected").addValueEventListener()
-
-            this.authorativeUserData = firebaseDatabase.getReference("authorative-user-storage");
-            this.clientRequests = firebaseDatabase.getReference("client-requests");
-            this.clientVisibleSubscriberRecords = firebaseDatabase.getReference("profiles");
-            this.recordsOfPurchase = firebaseDatabase.getReference("records-of-purchase");
-
-            // Used to listen in on new products from the Firebase product catalog.
-            this.quickBuyProducts = firebaseDatabase.getReference("quick-buy-products");
-            this.products = firebaseDatabase.getReference("products");
-        }
-
-        public void addProductCatalogListener(final ChildEventListener productCatalogListener) {
-            checkNotNull(productCatalogListener);
-            this.quickBuyProducts.addChildEventListener(productCatalogListener);
-            this.products.addChildEventListener(productCatalogListener);
-        }
-
-        public void addProductCatalogValueListener(ValueEventListener productCatalogValueEventListener) {
-            checkNotNull(productCatalogValueEventListener);
-            this.quickBuyProducts.addValueEventListener(productCatalogValueEventListener);
-            this.products.addValueEventListener(productCatalogValueEventListener);
-        }
-
-        public void addPurchaseEventListener(final ChildEventListener cel) {
-            checkNotNull(cel);
-            this.clientRequests.addChildEventListener(cel);
-        }
-
-        public String pushRecordOfPurchaseByMsisdn(Map<String, Object> asMap) {
-
-            final DatabaseReference dbref = recordsOfPurchase.push();
-
-            dbref.updateChildren(asMap);
-            return dbref.getKey();
-        }
-
-        public void updateClientVisibleUsageString(String msisdn, String gbLeft) throws StorageException {
-            final String key = getKeyFromPhoneNumber(clientVisibleSubscriberRecords, msisdn);
-            if (key == null) {
-                LOG.error("Could not find entry for phoneNumber = " + msisdn + " Not updating user visible storage");
-                return;
-            }
-
-            final Map<String, Object> displayRep = new HashMap<>();
-            displayRep.put("phoneNumber", msisdn);
-
-            displayRep.put("usage", gbLeft);
-
-            clientVisibleSubscriberRecords
-                    .child(key)
-                    .updateChildren(displayRep);
-        }
-
-        public void removeSubscriberByMsisdn(String msisdn)  throws StorageException{
-            removeByMsisdn(authorativeUserData, msisdn);
-        }
-
-        private void removeByMsisdn(
-                final DatabaseReference dbref,
-                final String msisdn) throws StorageException {
-            checkNotNull(msisdn);
-            checkNotNull(dbref);
-            final String key = getKeyFromMsisdn(dbref, msisdn);
-            if (key != null) {
-                removeChild(dbref, key);
-            }
-        }
-
-
-        private  String getKeyFromLookupKey(final DatabaseReference dbref, String msisdn, String lookupKey) throws StorageException {
-            final CountDownLatch cdl = new CountDownLatch(1);
-            final Set<String> result = new HashSet<>();
-            dbref.orderByChild(lookupKey)
-                    .equalTo(msisdn)
-                    .limitToFirst(1).addListenerForSingleValueEvent(new ValueEventListener() {
-                @Override
-                public void onDataChange(final DataSnapshot snapshot) {
-                    handleDataChange(snapshot, cdl, result, msisdn);
-                }
-
-                @Override
-                public void onCancelled(DatabaseError error) {
-
-                }
-            });
-
-            try {
-                if (!cdl.await(10, TimeUnit.SECONDS)) {
-                    throw new StorageException("Query timed out");
-                } else if (result.isEmpty()) {
-                    return null;
-                } else {
-                    return result.iterator().next();
-                }
-            } catch (InterruptedException e) {
-                throw new StorageException("Interrupted", e);
-            }
-        }
-
-        private  String getKeyFromPhoneNumber(final DatabaseReference dbref, final String msisdn) throws StorageException {
-            final String lookupKey = "phoneNumber";
-            return getKeyFromLookupKey(dbref, msisdn, lookupKey);
-        }
-
-        private  String getKeyFromMsisdn(final DatabaseReference dbref, final String msisdn) throws StorageException {
-            final String lookupKey = "msisdn";
-            return getKeyFromLookupKey(dbref, msisdn, lookupKey);
-        }
-
-        public void removeByMsisdn(String msisdn) throws StorageException {
-            removeByMsisdn(clientVisibleSubscriberRecords, msisdn);
-        }
-
-        public String injectPurchaseRequest(PurchaseRequest pr) {
-            final FbPurchaseRequest cr = (FbPurchaseRequest) pr;
-            final DatabaseReference dbref = clientRequests.push();
-            final Map<String, Object> crAsMap = cr.asMap();
-            dbref.setValue(crAsMap);
-            return dbref.getKey();
-        }
-
-        public void removeRecordOfPurchaseById(String id) {
-            removeChild(recordsOfPurchase, id);
-        }
-
-        private void removeChild(final DatabaseReference db, final String childId) {
-            // XXX Removes whole tree, not just the subtree for id.
-            //     how do I fix this?
-            checkNotNull(db);
-            checkNotNull(childId);
-            db.child(childId).getRef().removeValue();
-        }
-
-        public void removePurchaseRequestById(String id) {
-            checkNotNull(id);
-            removeChild(clientRequests, id);
-        }
-
-        public Subscriber getSubscriberFromMsisdn(String msisdn) throws StorageException {
-            final CountDownLatch cdl = new CountDownLatch(1);
-            final Set<Subscriber> result = new HashSet<>();
-
-            final Query q = authorativeUserData.orderByChild("msisdn").equalTo(msisdn).limitToFirst(1);
-
-            final ValueEventListener listenerThatWillReadSubcriberData = new ValueEventListener() {
-                @Override
-                public void onDataChange(final DataSnapshot snapshot) {
-
-                    if (!snapshot.hasChildren()) {
-                        cdl.countDown();
-                        return;
-                    } else {
-                        for (final DataSnapshot snap : snapshot.getChildren()) {
-
-                            final FbSubscriber sub =
-                                    snap.getValue(FbSubscriber.class);
-                            final String key = snap.getKey();
-
-                            sub.setFbKey(key);
-                            result.add(sub);
-                            cdl.countDown();
-                        }
-                    }
-                }
-
-                @Override
-                public void onCancelled(DatabaseError error) {
-                }
-            };
-
-            q.addListenerForSingleValueEvent(
-                    listenerThatWillReadSubcriberData);
-
-            try {
-                if (!cdl.await(10, TimeUnit.SECONDS)) {
-                    LOG.info("authorativeuserdata = '" + authorativeUserData
-                            + "', msisdn = '" + msisdn
-                            + "' => timeout");
-                    throw new StorageException("Query timed out. authorativeuserdata = '" + authorativeUserData + "', msisdn = '" + msisdn + "'");
-                } else if (result.isEmpty()) {
-                    LOG.info("authorativeuserdata = '" + authorativeUserData
-                            + "', msisdn = '" + msisdn
-                            + "' => null");
-                    return null;
-                } else {
-                    final Subscriber r = result.iterator().next();
-                    LOG.info("authorativeuserdata = '" + authorativeUserData + "', msisdn = '" + msisdn + "' => " + r);
-                    return r;
-                }
-            } catch (InterruptedException e) {
-                LOG.info("authorativeuserdata = '" + authorativeUserData + "', msisdn = '" + msisdn + "' => interrupted");
-                throw new StorageException("Interrupted", e);
-            }
-        }
-
-        public void updateAuthorativeUserData(FbSubscriber sub) {
-            final DatabaseReference dbref = authorativeUserData.child(sub.getFbKey());
-            dbref.updateChildren(sub.asMap());
-        }
-
-        public String insertNewSubscriber(FbSubscriber sub) {
-            final DatabaseReference dbref = authorativeUserData.push();
-            sub.setFbKey(dbref.getKey());
-            dbref.updateChildren(sub.asMap());
-            return dbref.getKey();
-        }
-
-        public Collection<Subscriber> getAllSubscribers() {
-            final Query q = authorativeUserData.orderByKey();
-            final Set<Subscriber> subscribers = new LinkedHashSet<>();
-            final CountDownLatch cdl = new CountDownLatch(1);
-            final ValueEventListener collectingVisitor =
-                    newListenerThatWillCollectAllSubscribers(subscribers, cdl);
-
-            q.addListenerForSingleValueEvent(collectingVisitor);
-
-            try {
-                cdl.await(10, TimeUnit.SECONDS);
-            } catch (InterruptedException e) {
-                LOG.error("Failed to get all subscribers");
-            }
-            if (cdl.getCount() > 0) {
-                LOG.error("Failed to get all subscribers");
-            }
-            return subscribers;
-        }
-
-        private  static ValueEventListener newListenerThatWillCollectAllSubscribers(final Set<Subscriber> subscribers, final CountDownLatch cdl) {
-            checkNotNull(subscribers);
-            checkNotNull(cdl);
-            return new ValueEventListener() {
-                @Override
-                public void onDataChange(final DataSnapshot snapshot) {
-                    if (!snapshot.hasChildren()) {
-                        cdl.countDown();
-                        return;
-                    }
-                    for (final DataSnapshot child : snapshot.getChildren()) {
-                        final FbSubscriber subscriber = child.getValue(FbSubscriber.class);
-                        subscriber.setFbKey(child.getKey());
-                        subscribers.add(subscriber);
-                    }
-                    cdl.countDown();
-                }
-                @Override
-                public void onCancelled(DatabaseError error) {
-                    LOG.error(error.getMessage(), error.toException());
-                }
-            };
-        }
-    }
 
 
     final FbDatabaseFacade facade;
@@ -578,7 +300,7 @@ public final class FbStorage implements Storage {
         facade.removePurchaseRequestById(id);
     }
 
-    private static void handleDataChange(final DataSnapshot snapshot, final CountDownLatch cdl, final Set<String> result, final String msisdn) {
+    static void handleDataChange(final DataSnapshot snapshot, final CountDownLatch cdl, final Set<String> result, final String msisdn) {
         if (!snapshot.hasChildren()) {
             cdl.countDown();
         } else try {
